==> Data Science Project/1 DemoGraphic_Data_Preparation.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report,confusion_matrix

## Demographic

df = pd.ExcelFile('Zip Code Recommendations Data.xlsx')

y_data = pd.read_excel(df,'InputData')

demographic = pd.read_csv("zip_codes_2018_5yr.csv")
demographic.head()

demographic.columns = ['zip']+list(demographic.columns[1:])
demographic = demographic.drop('do_date',axis=1)
demographic.head()

demographic.nunique()

demographic = demographic.dropna(axis=1,thresh=int(len(demographic)*0.5))
print(demographic.shape)
demographic.head()

demographic.shape

med_dict = dict(demographic.median())

demographic = demographic.fillna(value=med_dict,axis=0)
demographic.isnull().sum().sum()

zip_demographic = pd.DataFrame()
for i in y_data['ZipCode']:
    zip_demographic = pd.concat([zip_demographic,demographic[demographic['zip']==i]])
print(zip_demographic.shape)
zip_demographic.head()

correlation = abs(zip_demographic.corr())

corr = correlation.copy()
corr.head()

k = 0
for col in zip_demographic.columns[1:]:
    try:
        l = list(corr.columns[(corr.loc[col]>0.8) & (corr.loc[col]<1)])
        if len(l)!=0:
            corr = corr.drop(l,axis=1)
            corr = corr.drop(l,axis=0)
            print(f' {k+1} iteration removed columns are : {l}\n ')
            k+=1
    except:
        pass

corr.shape

zip_demographic = zip_demographic[corr.columns]
zip_demographic.head()

y_data.columns = ['zip','Liking']
y_data.head(5)

zip_demographic = zip_demographic.merge(y_data)
zip_demographic.head()

zip_demographic.shape

zip_demographic['Liking'].value_counts()

# zip_demographic.to_csv("Cleaned_DemoGraphic_data_2018.csv")
