==> Data Science Project/2 Data Preparation - MergedData.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns

import matplotlib.pyplot as plt

df = pd.ExcelFile('Zip Code Recommendations Data.xlsx')

df1 = pd.read_excel(df,'InputData')

## Based only on Apartment property
df1.head()

y_data = df1.copy()
y_data.head()

y_data['ZipCode']

## Extracting the avaliable data from all the sheets

df.sheet_names

## CurrentAttributesByZip

CurrentAttributes = pd.read_excel(df,'CurrentAttributesByZip')
print(CurrentAttributes.shape)
CurrentAttributes.head()

CurrentAttributes.columns = ['zip','sector']+list(CurrentAttributes.columns[2:])
CurrentAttributes.head()

## Collecting the zipcode data matching with the labeled data
zip_attribute = pd.DataFrame()
for i in y_data['ZipCode']:
     zip_attribute = pd.concat([zip_attribute,CurrentAttributes[CurrentAttributes['zip']==i]])
print(zip_attribute.shape)
zip_attribute.head()

## RentalTrendsBySubmarket

RentalTrendsBySubmarkets = pd.read_excel(df,"RentalTrendsBySubmarkets")
RentalTrendsBySubmarkets.head()

RentalTrendsBySubmarkets = RentalTrendsBySubmarkets.dropna(axis=1,thresh=int(len(RentalTrendsBySubmarkets)*0.5))
RentalTrendsBySubmarkets.head()

## Mapping Zip_submarket with rental trends

zip_to_submarket = pd.read_excel(df,'ZipToSubmarket')
zip_to_submarket.columns = ['zip','sector','market_name','submarket_name','number_of_properties']
zip_to_submarket.head()

## Collecting the zipcode data matching with the labeled data
zip_submarket_data = pd.DataFrame()
for i in y_data['ZipCode']:
     zip_submarket_data = pd.concat([zip_submarket_data,zip_to_submarket[zip_to_submarket['zip']==i]])
print(zip_submarket_data.shape)
zip_submarket_data.head()

zip_submarket_data = zip_submarket_data.merge(RentalTrendsBySubmarkets)
print(zip_submarket_data.shape)
zip_submarket_data.head()

## Rental Trends by market

rental_trends_by_market  = pd.read_excel(df,"RentalTrendsByMarket")
print(rental_trends_by_market.shape)
rental_trends_by_market.head()

rental_trends_by_market = rental_trends_by_market.dropna(axis=1,thresh=int(len(rental_trends_by_market)*0.5))
rental_trends_by_market.head()

## Mapping RentalTrends with ziptoMarket

zip_to_market = pd.read_excel(df,'ZipToMarket')
zip_to_market.head()

## Collecting the zipcode data matching with the labeled data
zip_market_data = pd.DataFrame()
for i in y_data['ZipCode']:
     zip_market_data = pd.concat([zip_market_data,zip_to_market[zip_to_market['zip']==i]])
zip_market_data.columns = ['zip','sector','market_name','number_of_properties']
print(zip_market_data.shape)
zip_market_data.head()

## Map

zip_market_data = pd.merge(zip_market_data,rental_trends_by_market)
print(zip_market_data.shape)
zip_market_data.head()

## ZipToCSA

zip_to_csa = pd.read_excel(df,'ZipToCSA')
zip_to_csa.columns = ['zip']+list(zip_to_csa.columns[1:])
zip_to_csa.head()

## Collecting the zipcode data matching with the labeled data
zip_csa = pd.DataFrame()
for i in y_data['ZipCode']:
     zip_csa = pd.concat([zip_csa,zip_to_csa[zip_to_csa['zip']==i]])
zip_csa.head()

## Collected Data

df_demo =  pd.read_csv('Cleaned_DemoGraphic_data_2018.csv')
df_demo = df_demo.drop(['Unnamed: 0','Liking'],axis=1)
df_demo.head()

l = [zip_attribute,
     zip_submarket_data,
     zip_market_data,
     zip_csa,
     df_demo]

l[1]['sector'].unique()

for i in l:
    print(i.shape)

for i in l:
    print(i.shape ,' : ', i.isnull().sum().sum())

# ## Combining market and submarket data
# market_submarket = l[1].merge(l[2],'outer')
# market_submarket.to_csv('Market_Submarket.csv')

# market_submarket.shape

merged = l[0].copy()
for i in l:
    merged = pd.merge(merged,i,how='outer')
merged['zip'] = merged['zip'].apply(int)
merged.shape

merged.head()

## Labeling Based on apartment and Houston

merged['market_name'].value_counts()

merged['sector'].value_counts()

label = df1.dropna(0)
label.columns = ['zip','Liking']
label['sector'] = 'Apartment'
label['market_name'] = 'Houston'
label.head()

label.shape,merged.shape

df1.columns,y_data.columns

df1.columns = ['zip','Liking']
y_data.columns = ['zip','Liking']

label = y_data.merge(label,how='outer')
print(label.shape)
print(label.dropna(0).shape)
label.head()

label.shape

merged['market_name'].unique()

merged_label = pd.merge(merged,label,how='outer')
print(merged_label.shape)
merged_label.head()

merged_label.Liking.value_counts()

# merged_label.to_csv('MergedData.csv')
