==> Data Science Project/3 Data EDA-Trainable_data.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

df = pd.read_csv('MergedData.csv')
df = df.drop('Unnamed: 0',axis=1)
df.head()

df.shape

df.shape,df['Liking'].value_counts()

label_data = df[~(df['Liking'].isnull())]
unlabel_data  = df[df['Liking'].isnull()]
print(label_data.shape,unlabel_data.shape)
label_data['Liking'].value_counts()

## Labeled Data EDA

label_data.head()

label_data.nunique()

for i in label_data:
    if label_data[i].nunique()<2:
        label_data = label_data.drop(i,axis=1)
print(label_data.shape)
label_data.head()

label_data = label_data.dropna(axis=1,thresh=int(len(label_data)*0.4))

label_data.shape

sum(label_data.dtypes=='object')

## Handeling missing value categorical and continous

for i in label_data:
    if label_data[i].dtypes=='O':
        print(i)
        print(label_data[i].value_counts())
        print()

label_data.shape,label_data.isnull().sum()

median_dict = dict(label_data.median())
median_dict['submarket_name'] = 'Montrose/River Oaks'
median_dict['building_class'] = 'All'
median_dict['period'] = 'Y'
median_dict

len(median_dict)

label_data_new = label_data.fillna(value=median_dict,axis=0)
print(label_data_new.isnull().sum().sum())
label_data_new.head()

# label_data_new.to_csv('RawLabeledData.csv')

## Data Preprocessing
- Label Encoding
- Min Max Scaling
- Standardise

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import StandardScaler

cleaned_labeled_data = label_data_new.copy()
categorical_dictonary = {}
for i in label_data_new.columns[1:-1]:
    if label_data_new[i].dtype =='O':
        le = LabelEncoder()
        cleaned_labeled_data[i] = le.fit_transform(label_data_new[i])
        categorical_dictonary[i] = dict(zip(le.classes_,list(range(len(le.classes_)))))
    else:
        min_max = StandardScaler()
        cleaned_labeled_data[i] = min_max.fit_transform(np.array(label_data_new[i]).reshape(-1,1))

categorical_dictonary

cleaned_labeled_data.head()

# cleaned_labeled_data.to_csv('CleanedLabeledData.csv')

## Standardisation

# test_data = stand_data[['zip','population','asking_rent_pct_chg','total_employment','industrial_employment','household_avg_income_pct_chg','Liking']]

# test_data.to_csv('test_data.csv')

## Corelation Check

corr = label_data_new.corr()

thresh = corr[(corr>0.8) | (corr<-0.8)]
thresh

high_corr_dict = {}
for i in thresh:
    l = thresh[i].dropna()
    if len(l)>1:
        print('Attribute Name :',i,'\n')
        print('High Correlations :\n')
        print(l)
        high_corr_dict[i]=l
        print()

high_corr_dict.keys()

len(label_data_new.columns)

## Removing corelated columns

temp = cleaned_labeled_data.drop(['zip','Liking'],axis=1)
corr = abs(temp.corr())
corr.shape

k = 0
for col in corr.columns[1:-1]:
    try:
        l = list(corr.columns[(corr.loc[col]>0.8) & (corr.loc[col]<1)])
        if len(l)!=0:
            corr = corr.drop(l,axis=1)
            corr = corr.drop(l,axis=0)
            print(f' {k+1} iteration removed columns are : {l}\n ')
            k+=1
    except:
        pass

corr.shape

test = cleaned_labeled_data['zip']
test = pd.concat([test,cleaned_labeled_data[corr.columns]],axis=1)
test['Liking'] = cleaned_labeled_data['Liking']
print(test.shape)
test.head()

test.Liking.value_counts()

# test.to_csv('Trainable_data.csv')
